Add a bunch of stuff
To be honest, I'm in a feature-adding craze, trying to get all this implemented to make it somewhat usable and testable. I still need to rigorously test the functions that I've written. I also need to ensure the number storage is done as I want it to be done (as unsigned 32-bit integers, but potentially as 16 or 24-bit numbers if the need arises).