Edit byte-array insertion/removal algorithms
Change `stripleadingzeros` to run through linearly, first moving elements then setting the rest to `nil`, as opposed to running through once simultaneously moving values and setting the old location to nil. Also edit `getbytearray` to return a copy of the parameter if it's a byte-array, meaning that it will return a byte-array of the same length: otherwise calculate a new byte-array. Add function `getminimizedbytearray` to return a byte-array of the shortest length that is a logical equalivalent to the passed byte-array.