Add trusted forms of table constructors
Add `constructorsignmagnitudetrusted` and `constructorbytearraytrusted` which do not create copies of the arrays being passed to them, hence they 'trust' the arrays being passed in. These are used for internal functions to mitigate frivolous table creations. Use with caution and only with arrays you know are already copies and will not cause damage.