Change `divideoneword` to use and return numbers instead of byte arrays
If a number is one-word long, then it doesn't need to go into an array. If the divisor is one-word long, then the remainder will be at-most one-word long; thus neither the divisor nor the remainder need to be arrays. This is to prevent making extra garbage tables that will ultimately not be used, as well as to make some math stuff easier (when using `divideoneword` internally.

Also, change the name to not include the word 'destructive' cos it isn't destructive