Copy minuend and strip leading zeros
Multiplication was broken for numbers with large groups of zeros in the middle/end because those zeros contributed to the length of the magnitude, making a really large Zero greater than a not as large positive number.\n\nThere was also an issue in subtraction when subtracting a smaller number by a larger number (as if calling subtractmagnitudes(a, b) if a < b). The function should work still by getting the absolute difference between the numbers regardless of order, but there was an issue in hot-mutating the minuend if the mutations would change the length of the minuend. The fix was to copy the minuend if the minuend was smaller (that is, when the length of the minuend would be changed).