Reformat code to follow some rules of mine
The rules are as follows (subject to change as I learn more about Lua):
* All variables are declared at the function level (like C)
* Constants are declared and defined before changing variables with a 1-line gap between
* Changing variables are not defined in the same line as their declaration
* There are to be two line gaps between blocks of code unless there is a direct progression from one line to another
* Line lengths are to be kept below 80 characters long (79 printing and 1 newline)
* Use a new `if` block after a return instead of an `elseif`
* Use the local references of library functions (`stringchar` instead of `string.char`)

The first three bullets are the most likely to change since they apparently could contribute to a bit of slow-down. When the Lua compiler compiles a chunk that declares a variable and assigns a value to it in two different lines it takes two different instructions: one to initialize it as `nil` and the other to assign the value. Thus, by declaring all the variables at the beginning I'm wasting a lot of time setting them all to `nil` just to redefine them later. Also, constant variables (such as the error message prefixes) could probably be moved to a static, global scope to prevent wasting precious clock cycles defining variables that may not be used. That's premature optimization, though. I'm a fan of declaring locals at the start of the functions, despite the Lua virtual machine taking two instructions to initialize it with its first value instead of one.