Fix leading zeros algorithm
It turns out Henry S Warren Jr specified multiple different algorithms to calculate the number of leading zeros in a 32-bit integer and I happened to pick the one that would be the slowest because of the function calls to the shifting functions. So, I changed all the left-shifts with multiplications and the right-shift-equals with comparisons.