Add "Word" and "Split-Long" to definitions
From now on, I will be distinguishing between 32-bit integers and other integers by using the terms "int" and "long". Specifically, an int is an unsigned 32-bit integer and nothing else. A long is a 64-bit integer that is often split into two ints, one for the upper word and one for the lower word.