Reformat doccomments
New formatting style rule: in doccomments, the first "paragraph" describes the function itself and uses the implicit subject. For example, instead of saying `This function returns a special number.` it says `Returns a special number.`. Any statement where the subject is not the function (implicitly or explicitly) will go into following paragraphs.