Allow `dest...mult..add` to work on non-allocated arrays
An allocated array is a word-array that is filled with zeros in every index from 1 to `#array`. Allocating arrays is useful if there will be random accesses into it when doing math (like multiplication), but if you're just adding to it like we are in string construction, then you don't need to allocate the array.