Don't create a copy of the minuend
I don't recall the logic behind making a copy of the minuend, but I don't think it's necessary. The result of the subtraction is put into the minuend at an index with respect to the longer length. The length of the minuend will be no longer than the longer length (which kind of fits the definition of the "longer length"), so regardless of if the target magnitude is the shorter or longer of the two, the writing of the words into it will not affect the calculation.