Replace `minuend` and `subtrahend`
with `thismag` and `thatmag`

The terms 'minuend' and 'subtrahend' imply an order of operands: subtraction is always `minuend - subtrahend`. This function, however, just calculates the absolute difference between them, which means there's no order. Thus, the use of the terms 'minuend' and 'subtrahend' are slightly misleading and as such I will replace them with the generic 'thismag' and 'thatmag'.