Rewrite the bitwise not functions
Bigintegers have a quicker method to performing a bitnot on them. As per the rules of Two's Complement, negation is a bitnot plus 1, so a bitnot is a negation minus 1. Flip the sign of the biginteger and move the magnitude closer to negative infinity (so if it turns out positive, decrement; negative, increment). This makes performing a bitnot on a biginteger a whole hell of a lot faster than getting the word-array (which involves copying the magnitude), bitnotting the whole word-array (which involves copying again), and getting a new sign and magnitude (which may involve copying a third time). Now it just increments or decrements in-place. Special care is taken in the case of 0 and -1.

Because the mutable function isn't using `bitwisenotwordarray`, put its procedure into `bitwisenot` and get rid of the function. This will make it marginally faster but more succinct.