Add a few ideas to to-do list
Number Namer is another project I built that can name any arbitrarily large number using Conway and Guy's large number naming scheme. The project is [here](https://harriknox.github.io/projects/numbernamer.html).

I think it would be neat to also include the ability to add separators (like the apostrophes in `12'345'678`) as well as the number of digits between them when printing the string-representation of the number. Adding that will break the reversibility of the function, however, as `constructorsignradix` cannot interpret strings with non-digits and non-signs in them. Attempting to do so complicates things.

Also, the public mutable functions should work on word-arrays as well, not just biginteger objects.