Add internal comparison functions
Add two functions `comparevalues` and `comparesignmagnitudes`. `comparevalues` is called by `equals` and `compare` after the values were verified by the operable value checker. This way `equals` isn't calling `compare`, which checks the operability of the values that were just checked for operability. `comparesignmagnitudes` compares the sign and magnitudes of trusted values so that we don't need to check the validity again and again and again.

This change was prompted by the fact that to check if two word-arrays were equal it would pass through both word arrays four times each before comparing:
1. in `equals` to check that the word-array is valid (checks every word in the array)
2. in `compare` to check that the word-array is valid again
3. in `getsignandmagnitude` to check the type to know which procedure to use
4. in `getwordarraysignandmagnitude` to copy the whole array

Now it passes through the word array only twice: once to check validity and to get the type, and the other to make a copy of the array for the magnitude.