Update `minimum`, `maximum`, and `minmax`
They look ugly now, but they're rewritten to use the new `comparesignmagnitudes`  and update `getsignandmagnitude` functions to be more efficient. Instead of doing two or four type-checks per loop, one of which is on a value that we already know is good, this just does the minimum number of type checks necessary.