Rewrite doccomment to fit the format of others
The general theme of the doccomments is that the first paragraph implicitly uses the function as the subject, allowing for likes like "Returns the number..." instead of "This function returns the number..." or "This returns the number...". I could have said "Uses one of the..." for the next sentence, but since it fits better with the last sentence, and the last sentence has a specific subject, I opted to leave it with the explicit "This".