Defer declaration of locality
Initially I declared all locals at the start of the function like in C, but that causes a few issues:
1. Declaration is usually separated from definition
2. It clutters the beginning of the function
3. I need to keep track in two places if I defined all locals I declared and declared all locals I defined

So, to make the code cleaner, I'm diverging from C and deferring the declarations of locality for all variables to right before they're used. The code looks nicer and it makes it more apparent when variables are first used. I did this for all of the functions I've already gone through. I will continue to do this for all later functions when I get to them.