Add work functions for adding/subbing magnitudes
Previously the `copyand(add|subtract)magnitudes` functions made a copy and then did the destructive operation on the copy, but that seemed inefficient and broke from the mold of `destructive...` and `copyand...` function pairs set by the `stripleadingzeros` and `negatewordarray` functions. In those functions there are the two entry functions with one work function. So I rewrote these functions to match that. There are two other sets of destructive-copy function pairs: increment and decrement, and left and right shift. The increment and decrement functions are probably going to stay untouched since the operation doesn't need to iterate through the whole array. The shifting functions are next.