Remove unnecessary cleanup
When you add two magnitudes you will never end up with leading zeros. By definition a magnitude cannot have leading zeros, so neither of the input magnitudes will have leading zeros. Thinking about this backwards, the only addition that would result in a leading zero would be an addition that would have caused an overflow (adding 0xFF and 1, for example). If such an overflow occurred, however, `carry` would be non-zero and would be prepended to the word-array, meaning there would be no leading zeros anyway. So it's an unnecessary function call.